[ConstraintSystem] NFC: Move `ConstraintSystem.h` to `include/swift/Sema`
