+++ lib/Sema/CSClosure.cpp
@@ -16,8 +16,8 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "ConstraintSystem.h"
 #include "TypeChecker.h"
+#include "swift/Sema/ConstraintSystem.h"
 
 using namespace swift;
 using namespace swift::constraints;
