+++ lib/Sema/CSStep.h
@@ -18,10 +18,10 @@
 #ifndef SWIFT_SEMA_CSSTEP_H
 #define SWIFT_SEMA_CSSTEP_H
 
-#include "ConstraintSystem.h"
 #include "swift/AST/Types.h"
 #include "swift/Sema/Constraint.h"
 #include "swift/Sema/ConstraintGraph.h"
+#include "swift/Sema/ConstraintSystem.h"
 #include "llvm/ADT/ArrayRef.h"
 #include "llvm/ADT/Optional.h"
 #include "llvm/ADT/STLExtras.h"
