+++ lib/Sema/IDETypeCheckingRequests.cpp
@@ -15,10 +15,10 @@
 #include "swift/AST/NameLookup.h"
 #include "swift/Basic/SourceManager.h"
 #include "swift/Frontend/Frontend.h"
+#include "swift/Sema/ConstraintSystem.h"
 #include "swift/Sema/IDETypeCheckingRequests.h"
 #include "swift/Subsystems.h"
 #include "TypeChecker.h"
-#include "ConstraintSystem.h"
 
 using namespace swift;
